refactor: 함수 정의 및 재사용 / 선언문 정리 (FE-131)
